+++ assets/app/a2services/patternmatching-service.js
@@ -13,25 +13,25 @@ angular.module('a2.srv.patternmatching', [
             return saveData;
         },
         list: function(callback) {
-            return $http.get('/api/project/'+Project.getUrl()+'/pattern_matchings').then(function(response){
+            return $http.get('/api/project/'+Project.getUrl()+'/pattern-matchings').then(function(response){
                 return response.data;
             }).catch(notify.serverError);
         },
         getDetailsFor: function(patternMatchingId) {
-            return $http.get('/api/project/' + Project.getUrl() + '/pattern_matchings/' + patternMatchingId + '/details').then(function(response){
+            return $http.get('/api/project/' + Project.getUrl() + '/pattern-matchings/' + patternMatchingId + '/details').then(function(response){
                 return response.data;
             });
         },
         getRoisFor: function(patternMatchingId, limit, offset) {
-            return $http.get('/api/project/' + Project.getUrl() + '/pattern_matchings/' + patternMatchingId + '/rois/' + (offset||0) + '_' + (limit||0)).then(function(response){
+            return $http.get('/api/project/' + Project.getUrl() + '/pattern-matchings/' + patternMatchingId + '/rois/' + (offset||0) + '_' + (limit||0)).then(function(response){
                 return response.data;
             });
         },
         getExportUrl: function(params){
-            return '/api/project/' + Project.getUrl() + '/pattern_matchings/' + params.patternMatching + '/rois.csv';
+            return '/api/project/' + Project.getUrl() + '/pattern-matchings/' + params.patternMatching + '/rois.csv';
         },
         validateRois: function(patternMatchingId, rois, validation) {
-            return $http.post('/api/project/' + Project.getUrl() + '/pattern_matchings/' + patternMatchingId + '/validate', {
+            return $http.post('/api/project/' + Project.getUrl() + '/pattern-matchings/' + patternMatchingId + '/validate', {
                 rois: rois,
                 validation: validation,
             }).then(function(response){
@@ -39,14 +39,14 @@ angular.module('a2.srv.patternmatching', [
             });
         },
         create: function(data) {
-            return $http.post('/api/project/'+Project.getUrl()+'/pattern_matchings/new', data).then(function(response){
+            return $http.post('/api/project/'+Project.getUrl()+'/pattern-matchings/new', data).then(function(response){
                 return response.data;
             });
         },
-        delete: function(classificationId, callback) {
-            // $http.get('/api/project/' + Project.getUrl() + '/classifications/' + classificationId + '/delete')
-            //     .success(callback)
-            //     .error(notify.serverError);
+        delete: function(patternMatchingId) {
+            return $http.post('/api/project/' + Project.getUrl() + '/pattern-matchings/' + patternMatchingId + '/remove').then(function(response){
+                return response.data;
+            });
         },
     };
 })